fix: revert kal with url
